\name{KRmisc-package}
\alias{KRmisc-package}
\alias{KRmisc}
\docType{package}
\title{
 Kornel's Miscellaneous R Functions
}
\description{
 Kornel's Miscellaneous R Functions
}
\details{
\tabular{ll}{
Package: \tab KRmisc\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Kornelius Rohmeyer

Maintainer: Kornelius Rohmeyer <rohmeyer@small-projects.de>
}
\references{
R Core Team (2013). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL http://www.R-project.org/.
}
\keyword{ package }
\seealso{
 \code{\link[gMCP:gMCP-package]{gMCP}} ~~
}
\examples{
# Nothing here yet.
}
